Tidy up xml parsing folder
